feat: divided code into multiple file, add new logo for program and also implemented class for hard drive encryption
