+++ utilities.py
@@ -1,0 +1,43 @@
+from PyQt5.QtGui import QColor
+import re
+
+
+class PasswordStrengthMeter:
+    @staticmethod
+    def calculate_strength(password):
+        if not password:
+            return 0
+
+        strength = 0
+
+        # Length check
+        length = len(password)
+        if length >= 8:
+            strength += 1
+        if length >= 12:
+            strength += 1
+        if length >= 16:
+            strength += 1
+
+        # Character diversity
+        if re.search(r"[A-Z]", password):
+            strength += 1
+        if re.search(r"[a-z]", password):
+            strength += 1
+        if re.search(r"[0-9]", password):
+            strength += 1
+        if re.search(r"[^A-Za-z0-9]", password):
+            strength += 1
+
+        # Normalize to 0-100 range
+        max_possible = 7  # 3 for length + 4 for diversity
+        return int((strength / max_possible) * 100)
+
+    @staticmethod
+    def get_strength_color(strength):
+        if strength < 30:
+            return QColor(255, 0, 0)  # Red
+        elif strength < 70:
+            return QColor(255, 255, 0)  # Yellow
+        else:
+            return QColor(0, 255, 0)  # Green
